+++ .gitignore
@@ -15,3 +15,4 @@
 local.properties
 .idea/
 app/release/
+.kotlin
