+++ app/src/main/java/com/slabstech/dhwani/voiceai/AuthenticatedActivity.kt
@@ -5,8 +5,12 @@ import android.os.Bundle
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
+import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.core.view.updatePadding
 import androidx.preference.PreferenceManager
 
 abstract class AuthenticatedActivity : AppCompatActivity() {
@@ -34,6 +38,39 @@ abstract class AuthenticatedActivity : AppCompatActivity() {
                 true
             }
             else -> super.onOptionsItemSelected(item)
+        }
+    }
+
+    /**
+     * For layouts with [R.id.coordinatorLayout], [R.id.bottomBar], [R.id.bottomNavigation]
+     * (Answer, Translate, Voice assistant): keeps content out of status bar, gesture nav, and IME.
+     */
+    protected fun setupAnswerStyleWindowInsets() {
+        val rootView = findViewById<View>(R.id.coordinatorLayout) ?: return
+        val bottomBar = findViewById<View>(R.id.bottomBar)
+        val bottomNav = findViewById<View>(R.id.bottomNavigation)
+
+        ViewCompat.setOnApplyWindowInsetsListener(rootView) { view, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            view.setPadding(0, systemBars.top, 0, 0)
+            insets
+        }
+
+        bottomBar?.let { bar ->
+            ViewCompat.setOnApplyWindowInsetsListener(bar) { view, insets ->
+                val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+                val imeInsets = insets.getInsets(WindowInsetsCompat.Type.ime())
+                view.updatePadding(bottom = imeInsets.bottom + systemBars.bottom + 8)
+                insets
+            }
+        }
+
+        bottomNav?.let { nav ->
+            ViewCompat.setOnApplyWindowInsetsListener(nav) { view, insets ->
+                val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+                view.updatePadding(bottom = systemBars.bottom)
+                insets
+            }
         }
     }
 }
