--- app/src/main/res/values/dimens.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- Clears labeled BottomNavigationView; FAB sits above it -->
-    <dimen name="voice_assistant_fab_margin_bottom">96dp</dimen>
-</resources>
